Review changes sugsested: Rm code duplication, Misleading Type Alias, Missing Documentation, Criterion 0.5, Unused Import, Sample Size Comment
