+++ prover/Cargo.toml
@@ -12,7 +12,7 @@ lazy_static = "1.5.0"
 
 [dev-dependencies]
 env_logger = "*"
-criterion = { version = "0.4", default-features = false }
+criterion = { version = "0.5", default-features = false }
 
 [[bench]]
 name = "vm_prover_benchmark"
